refactor: change Move class name and add squares of color methods
